+++ ml/context/initializers/initializers_test.go
@@ -14,12 +14,15 @@ import (
 func TestBroadcastTensorToShape(t *testing.T) {
 	graphtest.RunTestGraphFn(t, "BroadcastTensorToShape",
 		func(g *Graph) (inputs, outputs []*Node) {
-			t := tensors.FromValue([]float64{7.0, 11.0, 17.0})
-			inputs = []*Node{ConstCachedTensor(g, t)}
-			initFn := BroadcastTensorToShape(t)
+			t0 := tensors.FromValue([]float64{7.0, 11.0, 17.0})
+			t1 := tensors.FromScalar(3.0)
+			inputs = []*Node{ConstCachedTensor(g, t0), ConstCachedTensor(g, t1)}
+			initFn0 := BroadcastTensorToShape(t0)
+			initFn1 := BroadcastTensorToShape(t1)
 			outputs = []*Node{
-				initFn(g, shapes.Make(dtypes.Float32, 4, 3)),
-				initFn(g, shapes.Make(dtypes.Int32, 3)),
+				initFn0(g, shapes.Make(dtypes.Float32, 4, 3)),
+				initFn0(g, shapes.Make(dtypes.Int32, 3)),
+				initFn1(g, shapes.Make(dtypes.Int8, 1, 2, 3)),
 			}
 			return
 		}, []any{
@@ -30,5 +33,6 @@ func TestBroadcastTensorToShape(t *testing.T) {
 				{7.0, 11.0, 17.0},
 			},
 			[]int32{7, 11, 17},
+			[][][]int8{{{3, 3, 3}, {3, 3, 3}}},
 		}, 0)
 }